implementado validations nas entitys e intercionalização de mensagens…

+++ src/main/java/io/github/brunovicentealves/rest/dto/InformacaoItemPedidoDTO.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 public class InformacaoItemPedidoDTO {
 
     private String descricaoPedido;
-    private BigDecimal precoUnitario;
+    private Double precoUnitario;
     private Integer quantidade ;
 
 }
